+++ frontend/src/pages/GoogleRedirectPage/index.tsx
@@ -1,41 +1,14 @@
-import { useEffect, useContext } from "react";
-import { useLocation, useNavigate } from "react-router-dom";
-import { useMutation } from "@tanstack/react-query";
+import { useEffect } from "react";
+import { useLocation } from "react-router-dom";
 
-import { postGoogleOauth } from "@/api/googleOauth";
-import { UserContext } from "@/context/UserContext";
-
-import { PostGoogleOauthRequest } from "@/types/apiRequest";
+import useCreateGoogleOauthLogin from "./hooks/useCreateGoogleOauthLogin";
 
 const GoogleRedirectPage = () => {
-  const navigate = useNavigate();
-  const { login } = useContext(UserContext);
-
   const params = new URLSearchParams(useLocation().search);
   const authorizationCode = params.get("code");
   const inviteCode = params.get("state");
 
-  const { mutate: GoogleOauthLogin } = useMutation(
-    ({ authorizationCode, redirectUri }: PostGoogleOauthRequest) =>
-      postGoogleOauth({
-        authorizationCode,
-        redirectUri,
-      }),
-    {
-      onSuccess: (data) => {
-        if (data) {
-          login(data.accessToken, data.id);
-
-          if (inviteCode) {
-            navigate(`/invite/${inviteCode}`, { replace: true });
-            return;
-          }
-
-          navigate("/", { replace: true });
-        }
-      },
-    }
-  );
+  const { mutate: googleOauthLogin } = useCreateGoogleOauthLogin(inviteCode);
 
   useEffect(() => {
     const redirectUri = process.env.GOOGLE_REDIRECT_URL;
@@ -43,7 +16,7 @@ const GoogleRedirectPage = () => {
       return;
     }
 
-    GoogleOauthLogin({
+    googleOauthLogin({
       authorizationCode,
       redirectUri,
     });
